Move the classes that need to be hidden to the implementation package.
