+++ sdk/spring/spring-cloud-azure-stream-binder-eventhubs/src/main/java/com/azure/spring/cloud/stream/binder/eventhubs/EventHubsMessageChannelBinder.java
@@ -25,7 +25,7 @@ import com.azure.spring.integration.implementation.instrumentation.DefaultInstru
 import com.azure.spring.integration.implementation.instrumentation.DefaultInstrumentationManager;
 import com.azure.spring.integration.instrumentation.Instrumentation;
 import com.azure.spring.integration.instrumentation.InstrumentationManager;
-import com.azure.spring.integration.instrumentation.InstrumentationSendCallback;
+import com.azure.spring.integration.implementation.instrumentation.InstrumentationSendCallback;
 import com.azure.spring.messaging.ConsumerIdentifier;
 import com.azure.spring.messaging.ListenerMode;
 import com.azure.spring.messaging.PropertiesSupplier;
