+++ sdk/spring/spring-integration-azure-core/src/main/java/com/azure/spring/integration/implementation/instrumentation/AbstractProcessorInstrumentation.java
@@ -1,7 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.integration.instrumentation;
+package com.azure.spring.integration.implementation.instrumentation;
+
+import com.azure.spring.integration.instrumentation.Instrumentation;
 
 import java.time.Duration;
