+++ sdk/spring/spring-integration-azure-core/src/main/java/com/azure/spring/integration/implementation/instrumentation/InstrumentationSendCallback.java
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.integration.instrumentation;
+package com.azure.spring.integration.implementation.instrumentation;
 
+import com.azure.spring.integration.instrumentation.InstrumentationManager;
 import org.springframework.util.concurrent.ListenableFutureCallback;
 
 /**
