+++ sdk/spring/spring-integration-azure-core/src/test/java/com/azure/spring/integration/handler/DefaultMessageHandlerTests.java
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.integration.implementation.handler;
+package com.azure.spring.integration.handler;
 
-import com.azure.spring.integration.handler.DefaultMessageHandler;
 import com.azure.spring.messaging.AzureHeaders;
 import com.azure.spring.messaging.PartitionSupplier;
 import com.azure.spring.messaging.core.SendOperation;
