+++ sdk/spring/spring-integration-azure-eventhubs/src/main/java/com/azure/spring/integration/eventhubs/inbound/implementation/health/EventHubsProcessorInstrumentation.java
@@ -4,7 +4,7 @@
 package com.azure.spring.integration.eventhubs.inbound.implementation.health;
 
 import com.azure.messaging.eventhubs.models.ErrorContext;
-import com.azure.spring.integration.instrumentation.AbstractProcessorInstrumentation;
+import com.azure.spring.integration.implementation.instrumentation.AbstractProcessorInstrumentation;
 
 import java.time.Duration;
