+++ sdk/spring/spring-integration-azure-eventhubs/src/test/java/com/azure/spring/integration/eventhubs/inbound/implementation/health/EventHubsProcessorInstrumentationTests.java
@@ -5,7 +5,7 @@ package com.azure.spring.integration.eventhubs.inbound.implementation.health;
 
 import com.azure.messaging.eventhubs.models.ErrorContext;
 import com.azure.messaging.eventhubs.models.PartitionContext;
-import com.azure.spring.integration.instrumentation.AbstractProcessorInstrumentation;
+import com.azure.spring.integration.implementation.instrumentation.AbstractProcessorInstrumentation;
 import com.azure.spring.integration.implementation.instrumentation.AbstractProcessorInstrumentationTests;
 import com.azure.spring.integration.instrumentation.Instrumentation;
