+++ sdk/spring/spring-integration-azure-servicebus/src/main/java/com/azure/spring/integration/servicebus/inbound/implementation/health/ServiceBusProcessorInstrumentation.java
@@ -4,7 +4,7 @@
 package com.azure.spring.integration.servicebus.inbound.implementation.health;
 
 import com.azure.messaging.servicebus.ServiceBusErrorContext;
-import com.azure.spring.integration.instrumentation.AbstractProcessorInstrumentation;
+import com.azure.spring.integration.implementation.instrumentation.AbstractProcessorInstrumentation;
 
 import java.time.Duration;
