+++ sdk/spring/spring-integration-azure-servicebus/src/test/java/com/azure/spring/integration/servicebus/outbound/ServiceBusMessageHandlerTests.java
@@ -4,7 +4,7 @@
 package com.azure.spring.integration.servicebus.outbound;
 
 import com.azure.spring.integration.handler.DefaultMessageHandler;
-import com.azure.spring.integration.implementation.handler.DefaultMessageHandlerTests;
+import com.azure.spring.integration.handler.DefaultMessageHandlerTests;
 import com.azure.spring.messaging.PartitionSupplier;
 import com.azure.spring.servicebus.core.ServiceBusTemplate;
 import org.junit.jupiter.api.AfterEach;
